+++ moPepGen/cli/__init__.py
@@ -2,6 +2,7 @@
 from .generate_index import add_subparser_generate_index, generate_index
 from .call_variant_peptide import add_subparser_call_variant, call_variant_peptide
 from .call_noncoding_peptide import add_subparser_call_noncoding, call_noncoding_peptide
+from .call_alt_translation import add_subparser_call_alt_translation, call_alt_translation
 from .parse_vep import add_subparser_parse_vep, parse_vep
 from .parse_reditools import add_subparser_parse_reditools, parse_reditools
 from .parse_star_fusion import add_subparser_parse_star_fusion, parse_star_fusion