+++ test/integration/test_split_fasta.py
@@ -58,6 +58,7 @@ class TestSplitDatabase(TestCaseIntegration):
         ]
         args.variant_peptides = self.data_dir/'peptides/variant.fasta'
         args.noncoding_peptides = self.data_dir/'peptides/noncoding.fasta'
+        args.alt_translation_peptides = self.data_dir/'peptides/alt_translation.fasta'
         args.annotation_gtf = self.data_dir/'annotation.gtf'
         args.proteome_fasta = self.data_dir/'translate.fasta'
         cli.split_fasta(args)
@@ -80,6 +81,7 @@ class TestSplitDatabase(TestCaseIntegration):
         ]
         args.variant_peptides = self.data_dir/'peptides/variant.fasta'
         args.noncoding_peptides = self.data_dir/'peptides/noncoding.fasta'
+        args.alt_translation_peptides = self.data_dir/'peptides/alt_translation.fasta'
         args.annotation_gtf = self.data_dir/'annotation.gtf'
         args.proteome_fasta = self.data_dir/'translate.fasta'
         args.group_source = ['coding:gSNP,gINDEL']
@@ -102,6 +104,7 @@ class TestSplitDatabase(TestCaseIntegration):
         ]
         args.variant_peptides = self.data_dir/'peptides/variant.fasta'
         args.noncoding_peptides = self.data_dir/'peptides/noncoding.fasta'
+        args.alt_translation_peptides = self.data_dir/'peptides/alt_translation.fasta'
         args.annotation_gtf = self.data_dir/'annotation.gtf'
         args.proteome_fasta = self.data_dir/'translate.fasta'
         args.group_source = ['coding:gSNP,gINDEL']
@@ -125,6 +128,7 @@ class TestSplitDatabase(TestCaseIntegration):
         ]
         args.variant_peptides = self.data_dir/'peptides/variant.fasta'
         args.noncoding_peptides = self.data_dir/'peptides/noncoding.fasta'
+        args.alt_translation_peptides = self.data_dir/'peptides/alt_translation.fasta'
         args.annotation_gtf = self.data_dir/'annotation.gtf'
         args.proteome_fasta = self.data_dir/'translate.fasta'
         args.additional_split = ['Noncoding-gSNP']